+++ app/src/main/java/com/example/raycaster/View/Raycasting/Sprites/SpriteRenderer.java
@@ -1,11 +1,11 @@
 package com.example.raycaster.View.Raycasting.Sprites;
 
 public class SpriteRenderer {
-    public static void render(double heights,boolean texM,byte tex,int shadows){}
+    public static void render(float heights,boolean texM,byte tex,int shadows){}
 
-    public static void render(double heights,byte tex,int shadows){}
+    public static void render(float heights,byte tex,int shadows){}
 
-    public static void prepare(double heights, int tex,boolean texM){}
-    private static void prepare(double heights, int tex){}
+    public static void prepare(float heights, int tex,boolean texM){}
+    private static void prepare(float heights, int tex){}
 
 }
